pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghx"
version = "0.1.0"
description = "Intelligent GitHub CLI account switcher"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Todd Toler" },
]
dependencies = [
    "pyyaml>=6.0",
]
keywords = ["github", "cli", "account-switcher", "gh"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Version Control :: Git",
]

[project.scripts]
ghx = "ghx.cli:entry_point"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ghx"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
